legacy/ai_workforce_simulator/pyproject.toml
[project]
name = "ai-workforce-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloc>=0.2.5",
    "e2b-code-interpreter>=2.2.0",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.1",
    "langgraph-supervisor>=0.0.30",
    "python-dotenv>=1.1.1",
]
